Got the detector working (at least for the tests I've done)
Note: Right now the YOLO detector doesn't work because of versioning issues
between the pytorch in python 2 and python 3. There are also issues where
torch hub doesn't work because it creates temporary python files that
have non-ASCII characters (which aren't allowed in Python 2). I think the
path to solve this is to remove all references to torch hub and replace
this with a local model that has been modified to work with Python 2